[level 1] Title: 부족한 금액 계산하기, Time: 0.20 ms, Memory: 78.5 MB -BaekjoonHub

+++ 프로그래머스/1/82612. 부족한 금액 계산하기/부족한 금액 계산하기.java
@@ -3,7 +3,7 @@ class Solution {
         long payment = 0;
         long balance = 0;
         for (int i = 1; i <= count; i++) {
-            payment += price * i;
+            payment += (long) price * i;
         }
         
         balance = money - payment;
